Adding abstract class ConfigTableBaseComponent (fixes #1860)
