+++ AzureFunctions.AngularClient/src/app/site/site-config/app-settings/app-settings.component.ts
@@ -2,7 +2,7 @@ import { LogCategories } from './../../../shared/models/constants';
 import { Response } from '@angular/http';
 import { BroadcastService } from 'app/shared/services/broadcast.service';
 import { Component, Input, OnChanges, OnDestroy, SimpleChanges } from '@angular/core';
-import { FormArray, FormBuilder, FormGroup, Validators, ValidatorFn } from '@angular/forms';
+import { FormBuilder, FormGroup, Validators, ValidatorFn } from '@angular/forms';
 import { Observable } from 'rxjs/Observable';
 import { Subject } from 'rxjs/Subject';
 import { Subscription as RxSubscription } from 'rxjs/Subscription';
@@ -14,7 +14,7 @@ import { SaveOrValidationResult } from './../site-config.component';
 import { LogService } from './../../../shared/services/log.service';
 import { PortalResources } from './../../../shared/models/portal-resources';
 import { BusyStateScopeManager } from './../../../busy-state/busy-state-scope-manager';
-import { CustomFormControl, CustomFormGroup } from './../../../controls/click-to-edit/click-to-edit.component';
+import { CustomFormGroup } from './../../../controls/click-to-edit/click-to-edit.component';
 import { ArmObj, ArmObjMap } from './../../../shared/models/arm/arm-obj';
 import { CacheService } from './../../../shared/services/cache.service';
 import { AuthzService } from './../../../shared/services/authz.service';
@@ -24,14 +24,15 @@ import { RequiredValidator } from 'app/shared/validators/requiredValidator';
 import { LinuxAppSettingNameValidator } from 'app/shared/validators/linuxAppSettingNameValidator';
 import { ArmUtil } from 'app/shared/Utilities/arm-utils';
 
+import { ConfigTableBaseComponent }  from '../config-table-base-component';
+
 @Component({
   selector: 'app-settings',
   templateUrl: './app-settings.component.html',
   styleUrls: ['./../site-config.component.scss']
 })
-export class AppSettingsComponent implements OnChanges, OnDestroy {
+export class AppSettingsComponent extends ConfigTableBaseComponent implements OnChanges, OnDestroy {
   public Resources = PortalResources;
-  public groupArray: FormArray;
 
   private _resourceIdStream: Subject<string>;
   private _resourceIdSubscription: RxSubscription;
@@ -55,9 +56,6 @@ export class AppSettingsComponent implements OnChanges, OnDestroy {
   public loadingFailureMessage: string;
   public loadingMessage: string;
 
-  public newItem: CustomFormGroup;
-  public originalItemsDeleted: number;
-
   @Input() mainForm: FormGroup;
 
   @Input() resourceId: string;
@@ -71,12 +69,11 @@ export class AppSettingsComponent implements OnChanges, OnDestroy {
     private _authZService: AuthzService,
     broadcastService: BroadcastService
   ) {
+    super();
+
     this._busyManager = new BusyStateScopeManager(broadcastService, 'site-tabs');
 
     this._resetPermissionsAndLoadingState();
-
-    this.newItem = null;
-    this.originalItemsDeleted = 0;
 
     this._resourceIdStream = new Subject<string>();
     this._resourceIdSubscription = this._resourceIdStream
@@ -87,8 +84,7 @@ export class AppSettingsComponent implements OnChanges, OnDestroy {
         this._appSettingsArm = null;
         this._slotConfigNamesArm = null;
         this.groupArray = null;
-        this.newItem = null;
-        this.originalItemsDeleted = 0;
+        this._resetNewAndDeleted();
         this._resetPermissionsAndLoadingState();
         this._slotConfigNamesArmPath =
           `${new ArmSiteDescriptor(this.resourceId).getSiteOnlyResourceId()}/config/slotConfigNames`;
@@ -174,8 +170,7 @@ export class AppSettingsComponent implements OnChanges, OnDestroy {
   private _setupForm(appSettingsArm: ArmObj<any>, slotConfigNamesArm: ArmObj<SlotConfigNames>) {
     if (!!appSettingsArm && !!slotConfigNamesArm) {
       if (!this._saveError || !this.groupArray) {
-        this.newItem = null;
-        this.originalItemsDeleted = 0;
+        this._resetNewAndDeleted();
         this.groupArray = this._fb.array([]);
 
         this._requiredValidator = new RequiredValidator(this._translateService);
@@ -214,7 +209,7 @@ export class AppSettingsComponent implements OnChanges, OnDestroy {
           }
         }
 
-        this._validateAllControls(this.groupArray.controls as CustomFormGroup[]);
+        this._validateAllControls();
       }
 
       if (this.mainForm.contains('appSettings')) {
@@ -223,8 +218,7 @@ export class AppSettingsComponent implements OnChanges, OnDestroy {
         this.mainForm.addControl('appSettings', this.groupArray);
       }
     } else {
-      this.newItem = null;
-      this.originalItemsDeleted = 0;
+      this._resetNewAndDeleted();
       this.groupArray = null;
       if (this.mainForm.contains('appSettings')) {
         this.mainForm.removeControl('appSettings');
@@ -235,36 +229,14 @@ export class AppSettingsComponent implements OnChanges, OnDestroy {
   }
 
   validate(): SaveOrValidationResult {
-    const groups = this.groupArray.controls;
+    this._purgePristineNewItems();
 
-    // Purge any added entries that were never modified
-    for (let i = groups.length - 1; i >= 0; i--) {
-      const group = groups[i] as CustomFormGroup;
-      if (group.msStartInEditMode && group.pristine) {
-        groups.splice(i, 1);
-        if (group === this.newItem) {
-          this.newItem = null;
-        }
-      }
-    }
-
-    this._validateAllControls(groups as CustomFormGroup[]);
+    this._validateAllControls();
 
     return {
       success: this.groupArray.valid,
       error: this.groupArray.valid ? null : this._validationFailureMessage()
     };
-  }
-
-  private _validateAllControls(groups: CustomFormGroup[]) {
-    groups.forEach(group => {
-      const controls = (<FormGroup>group).controls;
-      for (const controlName in controls) {
-        const control = <CustomFormControl>controls[controlName];
-        control._msRunValidation = true;
-        control.updateValueAndValidity();
-      }
-    });
   }
 
   getConfigForSave(): ArmObjMap {
@@ -355,66 +327,8 @@ export class AppSettingsComponent implements OnChanges, OnDestroy {
     return this._translateService.instant(PortalResources.configUpdateFailureInvalidInput, { configGroupName: configGroupName });
   }
 
-  deleteItem(group: FormGroup) {
-    const groups = this.groupArray;
-    const index = groups.controls.indexOf(group);
-    if (index >= 0) {
-      if ((group as CustomFormGroup).msExistenceState === 'original') {
-        this._deleteOriginalItem(groups, group);
-      } else {
-        this._deleteAddedItem(groups, group, index);
-      }
-    }
-  }
-
-  private _deleteOriginalItem(groups: FormArray, group: FormGroup) {
-    // Keep the deleted group around with its state set to dirty.
-    // This keeps the overall state of this.groupArray and this.mainForm dirty.
-    group.markAsDirty();
-
-    // Set the group.msExistenceState to 'deleted' so we know to ignore it when validating and saving.
-    (group as CustomFormGroup).msExistenceState = 'deleted';
-
-    // Force the deleted group to have a valid state by clear all validators on the controls and then running validation.
-    for (const key in group.controls) {
-      const control = group.controls[key];
-      control.clearAsyncValidators();
-      control.clearValidators();
-      control.updateValueAndValidity();
-    }
-
-    this.originalItemsDeleted++;
-
-    groups.updateValueAndValidity();
-  }
-
-  private _deleteAddedItem(groups: FormArray, group: FormGroup, index: number) {
-    // Remove group from groups
-    groups.removeAt(index);
-    if (group === this.newItem) {
-      this.newItem = null;
-    }
-
-    // If group was dirty, then groups is also dirty.
-    // If all the remaining controls in groups are pristine, mark groups as pristine.
-    if (!group.pristine) {
-      let pristine = true;
-      for (const control of groups.controls) {
-        pristine = pristine && control.pristine;
-      }
-
-      if (pristine) {
-        groups.markAsPristine();
-      }
-    }
-
-    groups.updateValueAndValidity();
-  }
-
-  addItem() {
-    const groups = this.groupArray;
-
-    this.newItem = this._fb.group({
+  protected _createNewItem(): CustomFormGroup {
+    const newItem = this._fb.group({
       name: [
         null,
         Validators.compose(this._validatorFns)],
@@ -422,8 +336,6 @@ export class AppSettingsComponent implements OnChanges, OnDestroy {
       isSlotSetting: [false]
     }) as CustomFormGroup;
 
-    this.newItem.msExistenceState = 'new';
-    this.newItem.msStartInEditMode = true;
-    groups.push(this.newItem);
+    return newItem;
   }
 }
