+++ AzureFunctions.AngularClient/src/app/site/site-config/config-table-base-component.ts
@@ -1,0 +1,114 @@
+import { FormArray, FormGroup } from '@angular/forms';
+import { CustomFormControl, CustomFormGroup } from './../../controls/click-to-edit/click-to-edit.component';
+
+export abstract class ConfigTableBaseComponent {
+    public groupArray: FormArray;
+
+    public newItem: CustomFormGroup;
+    public originalItemsDeleted: number;
+
+    constructor() {
+        this._resetNewAndDeleted();
+    }
+
+    protected _resetNewAndDeleted() {
+        this.newItem = null;
+        this.originalItemsDeleted = 0;
+    }
+
+    protected _purgePristineNewItems() {
+        const groups = this.groupArray.controls;
+
+        // Purge any added entries that were never modified
+        for (let i = groups.length - 1; i >= 0; i--) {
+            const group = groups[i] as CustomFormGroup;
+            if (group.msStartInEditMode && group.pristine) {
+                groups.splice(i, 1);
+                if (group === this.newItem) {
+                    this.newItem = null;
+                }
+            }
+        }
+    }
+
+    protected _validateAllControls() {
+        const groups = this.groupArray.controls;
+
+        groups.forEach(group => {
+            const controls = (<FormGroup>group).controls;
+            for (const controlName in controls) {
+                const control = <CustomFormControl>controls[controlName];
+                control._msRunValidation = true;
+                control.updateValueAndValidity();
+            }
+        });
+    }
+
+    deleteItem(group: FormGroup) {
+        const groups = this.groupArray;
+        const index = groups.controls.indexOf(group);
+        if (index >= 0) {
+            if ((group as CustomFormGroup).msExistenceState === 'original') {
+                this._deleteOriginalItem(groups, group);
+            } else {
+                this._deleteAddedItem(groups, group, index);
+            }
+        }
+    }
+
+    protected _deleteOriginalItem(groups: FormArray, group: FormGroup) {
+        // Keep the deleted group around with its state set to dirty.
+        // This keeps the overall state of this.groupArray and this.mainForm dirty.
+        group.markAsDirty();
+
+        // Set the group.msExistenceState to 'deleted' so we know to ignore it when validating and saving.
+        (group as CustomFormGroup).msExistenceState = 'deleted';
+
+        // Force the deleted group to have a valid state by clear all validators on the controls and then running validation.
+        for (const key in group.controls) {
+            const control = group.controls[key];
+            control.clearAsyncValidators();
+            control.clearValidators();
+            control.updateValueAndValidity();
+        }
+
+        this.originalItemsDeleted++;
+
+        groups.updateValueAndValidity();
+    }
+
+    protected _deleteAddedItem(groups: FormArray, group: FormGroup, index: number) {
+        // Remove group from groups
+        groups.removeAt(index);
+        if (group === this.newItem) {
+            this.newItem = null;
+        }
+
+        // If group was dirty, then groups is also dirty.
+        // If all the remaining controls in groups are pristine, mark groups as pristine.
+        if (!group.pristine) {
+            let pristine = true;
+            for (const control of groups.controls) {
+                pristine = pristine && control.pristine;
+            }
+
+            if (pristine) {
+                groups.markAsPristine();
+            }
+        }
+
+        groups.updateValueAndValidity();
+    }
+
+    addItem() {
+        const groups = this.groupArray;
+
+        this.newItem = this._createNewItem();
+        this.newItem.msExistenceState = 'new';
+        this.newItem.msStartInEditMode = true;
+
+        groups.push(this.newItem);
+    }
+
+    protected abstract _createNewItem(): CustomFormGroup;
+}
