+++ AzureFunctions.AngularClient/src/app/site/site-config/connection-strings/connection-strings.component.ts
@@ -2,7 +2,7 @@ import { LogCategories } from './../../../shared/models/constants';
 import { Response } from '@angular/http';
 import { BroadcastService } from './../../../shared/services/broadcast.service';
 import { Component, Input, OnChanges, OnDestroy, SimpleChanges } from '@angular/core';
-import { FormArray, FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { Observable } from 'rxjs/Observable';
 import { Subject } from 'rxjs/Subject';
 import { Subscription as RxSubscription } from 'rxjs/Subscription';
@@ -16,7 +16,7 @@ import { LogService } from './../../../shared/services/log.service';
 import { PortalResources } from './../../../shared/models/portal-resources';
 import { DropDownElement } from './../../../shared/models/drop-down-element';
 import { BusyStateScopeManager } from './../../../busy-state/busy-state-scope-manager';
-import { CustomFormControl, CustomFormGroup } from './../../../controls/click-to-edit/click-to-edit.component';
+import { CustomFormGroup } from './../../../controls/click-to-edit/click-to-edit.component';
 import { ArmObj, ArmObjMap } from './../../../shared/models/arm/arm-obj';
 import { CacheService } from './../../../shared/services/cache.service';
 import { AuthzService } from './../../../shared/services/authz.service';
@@ -24,14 +24,15 @@ import { ArmSiteDescriptor } from 'app/shared/resourceDescriptors';
 import { UniqueValidator } from 'app/shared/validators/uniqueValidator';
 import { RequiredValidator } from 'app/shared/validators/requiredValidator';
 
+import { ConfigTableBaseComponent }  from '../config-table-base-component';
+
 @Component({
     selector: 'connection-strings',
     templateUrl: './connection-strings.component.html',
     styleUrls: ['./../site-config.component.scss']
 })
-export class ConnectionStringsComponent implements OnChanges, OnDestroy {
+export class ConnectionStringsComponent extends ConfigTableBaseComponent implements OnChanges, OnDestroy {
     public Resources = PortalResources;
-    public groupArray: FormArray;
 
     private _resourceIdStream: Subject<string>;
     private _resourceIdSubscription: RxSubscription;
@@ -53,9 +54,6 @@ export class ConnectionStringsComponent implements OnChanges, OnDestroy {
     public loadingFailureMessage: string;
     public loadingMessage: string;
 
-    public newItem: CustomFormGroup;
-    public originalItemsDeleted: number;
-
     @Input() mainForm: FormGroup;
 
     @Input() resourceId: string;
@@ -69,12 +67,11 @@ export class ConnectionStringsComponent implements OnChanges, OnDestroy {
         private _authZService: AuthzService,
         broadcastService: BroadcastService
     ) {
+        super();
+
         this._busyManager = new BusyStateScopeManager(broadcastService, 'site-tabs');
 
         this._resetPermissionsAndLoadingState();
-
-        this.newItem = null;
-        this.originalItemsDeleted = 0;
 
         this._resourceIdStream = new Subject<string>();
         this._resourceIdSubscription = this._resourceIdStream
@@ -85,8 +82,7 @@ export class ConnectionStringsComponent implements OnChanges, OnDestroy {
                 this._connectionStringsArm = null;
                 this._slotConfigNamesArm = null;
                 this.groupArray = null;
-                this.newItem = null;
-                this.originalItemsDeleted = 0;
+                this._resetNewAndDeleted();
                 this._resetPermissionsAndLoadingState();
                 this._slotConfigNamesArmPath =
                     `${new ArmSiteDescriptor(this.resourceId).getSiteOnlyResourceId()}/config/slotConfigNames`;
@@ -166,8 +162,7 @@ export class ConnectionStringsComponent implements OnChanges, OnDestroy {
     private _setupForm(connectionStringsArm: ArmObj<ConnectionStrings>, slotConfigNamesArm: ArmObj<SlotConfigNames>) {
         if (!!connectionStringsArm && !!slotConfigNamesArm) {
             if (!this._saveError || !this.groupArray) {
-                this.newItem = null;
-                this.originalItemsDeleted = 0;
+                this._resetNewAndDeleted();
                 this.groupArray = this._fb.array([]);
 
                 this._requiredValidator = new RequiredValidator(this._translateService);
@@ -205,7 +200,7 @@ export class ConnectionStringsComponent implements OnChanges, OnDestroy {
                     }
                 }
 
-                this._validateAllControls(this.groupArray.controls as CustomFormGroup[]);
+                this._validateAllControls();
             }
 
             if (this.mainForm.contains('connectionStrings')) {
@@ -214,8 +209,7 @@ export class ConnectionStringsComponent implements OnChanges, OnDestroy {
                 this.mainForm.addControl('connectionStrings', this.groupArray);
             }
         } else {
-            this.newItem = null;
-            this.originalItemsDeleted = 0;
+            this._resetNewAndDeleted();
             this.groupArray = null;
             if (this.mainForm.contains('connectionStrings')) {
                 this.mainForm.removeControl('connectionStrings');
@@ -226,36 +220,14 @@ export class ConnectionStringsComponent implements OnChanges, OnDestroy {
     }
 
     validate(): SaveOrValidationResult {
-        const groups = this.groupArray.controls;
+        this._purgePristineNewItems();
 
-        // Purge any added entries that were never modified
-        for (let i = groups.length - 1; i >= 0; i--) {
-            const group = groups[i] as CustomFormGroup;
-            if (group.msStartInEditMode && group.pristine) {
-                groups.splice(i, 1);
-                if (group === this.newItem) {
-                    this.newItem = null;
-                }
-            }
-        }
-
-        this._validateAllControls(groups as CustomFormGroup[]);
+        this._validateAllControls();
 
         return {
             success: this.groupArray.valid,
             error: this.groupArray.valid ? null : this._validationFailureMessage()
         };
-    }
-
-    private _validateAllControls(groups: CustomFormGroup[]) {
-        groups.forEach(group => {
-            const controls = (<FormGroup>group).controls;
-            for (const controlName in controls) {
-                const control = <CustomFormControl>controls[controlName];
-                control._msRunValidation = true;
-                control.updateValueAndValidity();
-            }
-        });
     }
 
     getConfigForSave(): ArmObjMap {
@@ -352,67 +324,10 @@ export class ConnectionStringsComponent implements OnChanges, OnDestroy {
         return this._translateService.instant(PortalResources.configUpdateFailureInvalidInput, { configGroupName: configGroupName });
     }
 
-    deleteItem(group: FormGroup) {
-        const groups = this.groupArray;
-        const index = groups.controls.indexOf(group);
-        if (index >= 0) {
-            if ((group as CustomFormGroup).msExistenceState === 'original') {
-                this._deleteOriginalItem(groups, group);
-            } else {
-                this._deleteAddedItem(groups, group, index);
-            }
-        }
-    }
-
-    private _deleteOriginalItem(groups: FormArray, group: FormGroup) {
-        // Keep the deleted group around with its state set to dirty.
-        // This keeps the overall state of this.groupArray and this.mainForm dirty.
-        group.markAsDirty();
-
-        // Set the group.msExistenceState to 'deleted' so we know to ignore it when validating and saving.
-        (group as CustomFormGroup).msExistenceState = 'deleted';
-
-        // Force the deleted group to have a valid state by clear all validators on the controls and then running validation.
-        for (const key in group.controls) {
-            const control = group.controls[key];
-            control.clearAsyncValidators();
-            control.clearValidators();
-            control.updateValueAndValidity();
-        }
-
-        this.originalItemsDeleted++;
-
-        groups.updateValueAndValidity();
-    }
-
-    private _deleteAddedItem(groups: FormArray, group: FormGroup, index: number) {
-        // Remove group from groups
-        groups.removeAt(index);
-        if (group === this.newItem) {
-            this.newItem = null;
-        }
-
-        // If group was dirty, then groups is also dirty.
-        // If all the remaining controls in groups are pristine, mark groups as pristine.
-        if (!group.pristine) {
-            let pristine = true;
-            for (const control of groups.controls) {
-                pristine = pristine && control.pristine;
-            }
-
-            if (pristine) {
-                groups.markAsPristine();
-            }
-        }
-
-        groups.updateValueAndValidity();
-    }
-
-    addItem() {
-        const groups = this.groupArray;
+    protected _createNewItem(): CustomFormGroup {
         const connectionStringDropDownTypes = this._getConnectionStringTypes(ConnectionStringType.SQLAzure);
 
-        this.newItem = this._fb.group({
+        const newItem = this._fb.group({
             name: [
                 null,
                 Validators.compose([
@@ -425,11 +340,9 @@ export class ConnectionStringsComponent implements OnChanges, OnDestroy {
             isSlotSetting: [false]
         }) as CustomFormGroup;
 
-        (<any>this.newItem).csTypes = connectionStringDropDownTypes;
+        (<any>newItem).csTypes = connectionStringDropDownTypes;
 
-        this.newItem.msExistenceState = 'new';
-        this.newItem.msStartInEditMode = true;
-        groups.push(this.newItem);
+        return newItem;
     }
 
     private _getConnectionStringTypes(defaultType: ConnectionStringType) {
